+++ terminal.c
@@ -2255,12 +2255,24 @@ term_reset_view(struct terminal *term)
 }
 
 void
+term_save_cursor(struct terminal *term)
+{
+    term->grid->saved_cursor = term->grid->cursor;
+    term->vt.saved_attrs = term->vt.attrs;
+    term->saved_charsets = term->charsets;
+}
+
+void
 term_restore_cursor(struct terminal *term, const struct cursor *cursor)
 {
     int row = min(cursor->point.row, term->rows - 1);
     int col = min(cursor->point.col, term->cols - 1);
+
     term_cursor_to(term, row, col);
     term->grid->cursor.lcf = cursor->lcf;
+
+    term->vt.attrs = term->vt.saved_attrs;
+    term->charsets = term->saved_charsets;
 }
 
 void